\name{plot3D.rgl-package}
\alias{plot3D.rgl-package}
\alias{plot3D.rgl}
\docType{package}
\title{
  Functions to plot multi-dimensional data using \code{rgl}.
}
\description{
  Package \code{plot3D.rgl} provides an interface between function from package
  \code{plot3D} and package \code{rgl}.

  It will plot most (but not all) features from plots generated with \code{plot3D}, except
  the color keys and polygons. Scatter plots will be represented by spherical
  points.  
}

\author{
  Karline Soetaert
}
%\references{
%% to be filled
%}
\seealso{
Functions from R-package \code{plot3D}.

Functions from R-package \code{rgl}.

\link{plotrgl} the only function in the package.
}


\keyword{ package }
